add kinematic_solver.py with kinematic fcns
